+++ CHANGELOG.md
@@ -19,6 +19,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 
+## [0.8.2] - unreleased
+
+### Fixed
+- initial SystemV start script. 
+
 ## [0.8.1] - unreleased
 
 ### Added